MV github api functions to seperate module
